requirements.txt
ipdb
wandb
jieba
nltk
rouge
pytorch-lightning
transformers